Flytte IInntektsjusteringForm til samme fil som andre interfacer
