+++ apps/selvbetjening-ui/src/inntektsjustering/Inntektsjustering.tsx
@@ -3,7 +3,7 @@ import {Route, Routes} from "react-router-dom";
 import React, {useEffect, useState} from "react";
 import {Heading, Label} from "@navikt/ds-react";
 import {getInntektsjustering} from "~api/api";
-import {IInntektsjustering} from "~inntektsjustering/InntektsjusteringDto";
+import {IInntektsjustering, IInntektsjusteringForm} from "~inntektsjustering/InntektsjusteringDto";
 import Navigation from "~components/Navigation";
 import {formaterDatoStrengTilLocaleDateTime} from "~utils/dates";
 import {FormProvider, useForm} from "react-hook-form";
@@ -14,18 +14,6 @@ export const inntektsjusteringPath = {
     form: '/inntektsjustering/inntekt',
     oppsummering: '/inntektsjustering/oppsummering',
     kvittering: '/inntektsjustering/kvittering',
-}
-
-export interface IInntektsjusteringForm {
-    skalHaArbeidsinntekt: boolean
-    skalHaNaeringsinntekt: boolean
-    skalHaInntektNorge: boolean
-    skalHaInntektUtland: boolean
-
-    arbeidsinntekt?: number
-    naeringsinntekt?: number
-    arbeidsinntektUtland?: number
-    naeringsinntektUtland?: number
 }
 
 export default function Inntektsjustering() {
