+++ apps/selvbetjening-ui/src/inntektsjustering/InntektsjusteringDto.ts
@@ -13,3 +13,14 @@ export interface IInntektsjusteringLagre {
     naeringsinntektUtland: number
 }
 
+export interface IInntektsjusteringForm {
+    skalHaArbeidsinntekt: boolean
+    skalHaNaeringsinntekt: boolean
+    skalHaInntektNorge: boolean
+    skalHaInntektUtland: boolean
+
+    arbeidsinntekt?: number
+    naeringsinntekt?: number
+    arbeidsinntektUtland?: number
+    naeringsinntektUtland?: number
+}
