+++ apps/selvbetjening-ui/src/inntektsjustering/InntektsjusteringForm.tsx
@@ -11,7 +11,8 @@ import styled from "styled-components";
 import React from "react";
 import {RHFCheckbox} from "~common/rhc/RHFCheckbox";
 import Navigation from "~components/Navigation";
-import {IInntektsjusteringForm, inntektsjusteringPath} from "~inntektsjustering/Inntektsjustering";
+import {inntektsjusteringPath} from "~inntektsjustering/Inntektsjustering";
+import {IInntektsjusteringForm} from "~inntektsjustering/InntektsjusteringDto";
 
 export default function InntektsjusteringForm({}) {
     const {t} = useTranslation('inntektsjustering')
